[Bug修复](dev_2.2.2): 修复bug
查询治疗计划列表和详情时，增加返回参数orgId和patientId

+++ yunya-modules/yunya-emr/src/main/java/com/yunya/modules/emr/biz/TreatPlanRecordBiz.java
@@ -347,7 +347,7 @@ public class TreatPlanRecordBiz extends BaseBiz<TreatPlanRecordMapper, TreatPlan
                 TreatPlanRecordInfoVO vo = new TreatPlanRecordInfoVO();
                 vo.setPlanId(entity.getId());
                 vo.setPlanName(entity.getPlanName());
-                vo.setStatus(entity.getStatus());
+                vo.setStatus(entity.getStatus().intValue());
                 vo.setCrtTime(entity.getCrtTime());
                 OrganizationInfo org = remoteSystemServiceFeign.findOrgInfoByOrgId(entity.getOrgId());
                 if (!ObjectUtils.isEmpty(org)) {
